Separate avatar upload into its own component
