--- src/components/Avatar/Avatar.scss
@@ -1,28 +1,0 @@
-@use '../../utilities/colors' as v;
-
-.avatar {
-  height: 60px;
-  width: 60px;
-}
-
-.avatar-input {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-
-  input::-webkit-file-upload-button {
-    border-radius: 1rem;
-    padding: 6px 10px;
-    background: #FFFFFF;
-    color: v.$green;
-    font-weight: 600;
-    font-size: 0.875rem;
-    border: 2px solid v.$green;
-  }
-}
-
-.img-input {
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-}
