+++ src/utilities/_avatars.scss
@@ -2,5 +2,5 @@
 .avatar {
   height: 45px;
   width: 45px;
-  border-radius: 100%;
+  border-radius: 50%;
 }
